Move Tag object to own module
