+++ radish/model.py
@@ -18,14 +18,6 @@ class Model(object):
     """
         Represents a base model
     """
-    class Tag(object):
-        """
-            Represents a tag for a model
-        """
-        def __init__(self, name, arg=None):
-            self.name = name
-            self.arg = arg
-
     class Context(object):
         """
             Represents a Models context.
