+++ tests/unit/test_feature.py
@@ -8,6 +8,7 @@ from radish.scenariooutline import ScenarioOutline
 from radish.scenarioloop import ScenarioLoop
 from radish.stepmodel import Step
 from radish.main import setup_config
+from radish.tags import Tag
 
 
 class FeatureTestCase(RadishTestCase):
@@ -131,7 +132,7 @@ class FeatureTestCase(RadishTestCase):
         """
             Test feature's has to run functionality
         """
-        f = Feature(1, "Feature", "Some feature", None, None, [Feature.Tag("foo", None), Feature.Tag("bar", None), Feature.Tag("bad_case", None)])
+        f = Feature(1, "Feature", "Some feature", None, None, [Tag("foo", None), Tag("bar", None), Tag("bad_case", None)])
         f.has_to_run.when.called_with(None, ["foo"], None).should.return_value(True)
         f.has_to_run.when.called_with(None, ["good_case", "foo"], None).should.return_value(True)
         f.has_to_run.when.called_with(None, ["good_case", "bar", "bad_case"], None).should.return_value(True)
